# health_service/[OLD]health_interface/swagger_server/models/bad_response.py
# coding: utf-8

from __future__ import absolute_import
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from swagger_server.models.base_model_ import Model
from swagger_server import util


class BadResponse(Model):
    """NOTE: This class is auto generated by the swagger code generator program.

    Do not edit the class manually.
    """
    def __init__(self, command: str=None, type: str=None, description: str=None):  # noqa: E501
        """BadResponse - a model defined in Swagger

        :param command: The command of this BadResponse.  # noqa: E501
        :type command: str
        :param type: The type of this BadResponse.  # noqa: E501
        :type type: str
        :param description: The description of this BadResponse.  # noqa: E501
        :type description: str
        """
        self.swagger_types = {
            'command': str,
            'type': str,
            'description': str
        }

        self.attribute_map = {
            'command': 'command',
            'type': 'type',
            'description': 'description'
        }
        self._command = command
        self._type = type
        self._description = description

    @classmethod
    def from_dict(cls, dikt) -> 'BadResponse':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The BadResponse of this BadResponse.  # noqa: E501
        :rtype: BadResponse
        """
        return util.deserialize_model(dikt, cls)

    @property
    def command(self) -> str:
        """Gets the command of this BadResponse.

        type of results: ok or error  # noqa: E501

        :return: The command of this BadResponse.
        :rtype: str
        """
        return self._command

    @command.setter
    def command(self, command: str):
        """Sets the command of this BadResponse.

        type of results: ok or error  # noqa: E501

        :param command: The command of this BadResponse.
        :type command: str
        """

        self._command = command

    @property
    def type(self) -> str:
        """Gets the type of this BadResponse.

        short description of the error  # noqa: E501

        :return: The type of this BadResponse.
        :rtype: str
        """
        return self._type

    @type.setter
    def type(self, type: str):
        """Sets the type of this BadResponse.

        short description of the error  # noqa: E501

        :param type: The type of this BadResponse.
        :type type: str
        """

        self._type = type

    @property
    def description(self) -> str:
        """Gets the description of this BadResponse.

        Brief description of the operation result  # noqa: E501

        :return: The description of this BadResponse.
        :rtype: str
        """
        return self._description

    @description.setter
    def description(self, description: str):
        """Sets the description of this BadResponse.

        Brief description of the operation result  # noqa: E501

        :param description: The description of this BadResponse.
        :type description: str
        """

        self._description = description