# health_service/[OLD]health_interface/swagger_server/swagger/swagger.yaml
openapi: 3.0.0
info:
  title: API
  description: "Interface for communicating with the Health-Monitor service. Its functionalities\
    \ can be organized in the following set of operations:\n   - Operations for service\
    \ management:  permits to add and remove docker hosts to the system and to uninstall\
    \ it\n   - Operations for container management: permits to see the managed containers,\
    \ their status and to change the container manager configuration\n   - Operations\
    \ for testing purpose: permits to test the service with configurable attacks to\
    \ the containers"
  contact:
    name: Lost(In The Cloud)
  version: 1.0.0
servers:
- url: http://172.16.3.167:8080/healthmonitor
  description: Unipi Virtual Machine offered by the course. A VPN connection to the
    unipi cloud is required
paths:
  /containers:
    get:
      summary: Gives all the managed containers by the service
      description: It doesn't require any parameter and gives back a set of lists
        each one describing the containers inside a particular managed docker host.
        The response will contain for each docker host its status and a list of all
        the containers with information about their status and the current packet
        loss measured
      operationId: get_all_containers
      parameters: []
      responses:
        "200":
          description: Correct response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CorrectResponse'
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadResponse'
      x-openapi-router-controller: swagger_server.controllers.default_controller
    post:
      summary: 'Change the threshold configured into each manager. '
      description: Change the threshold configured into the docker managers. Each
        manager have a personal threshold used to identify if the packet loss percentage
        is too high and the container needs to be restarted.
      operationId: change_all_thresholds
      parameters: []
      requestBody:
        description: The request requires the threshold value to be set into the managers.
          The value is a percentage so it needs to be between 0 and 100 to be accepted.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/containers_body'
        required: true
      responses:
        "200":
          description: Correct Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CorrectResponse'
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadResponse'
      x-openapi-router-controller: swagger_server.controllers.default_controller
  /containers/{address}:
    get:
      summary: Get the status of the selected docker host and a list of all the managed
        containers with their status and the current packet loss percentage measured
      description: It doesn't require any parameter and gives back a set of lists
        each one describing the containers inside a particular managed docker host.
        The response will contain for each docker host its status and a list of all
        the containers with information about their status and the current packet
        loss measured
      operationId: get_manager_containers
      parameters:
      - name: address
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: string
          description: The IPv4 address of the selected machine
          example: 172.16.3.167
      responses:
        "200":
          description: Correct Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CorrectResponse'
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadResponse'
      x-openapi-router-controller: swagger_server.controllers.default_controller
    put:
      summary: Add a new docker host into the service
      description: Install a new manager into the the selected machine. The operations
        requires the root password of the destination machine and that a ssh connection
        is present on the default port 23
      operationId: add_docker_manager
      parameters:
      - name: address
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: string
          description: The IPv4 address of the selected machine
          example: 172.16.3.167
      requestBody:
        description: The operation requires the password of the root user of the destination
          machine
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/containers_address_body'
        required: true
      responses:
        "200":
          description: Correct Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CorrectResponse'
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadResponse'
      x-openapi-router-controller: swagger_server.controllers.default_controller
    post:
      summary: 'Change the threshold configured into the selected manager. '
      description: Change the threshold configured into the selected docker manager.
        Each manager have a personal threshold used to identify if the packet loss
        percentage is too high and the container needs to be restarted.
      operationId: change_manager_threshold
      parameters:
      - name: address
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: string
          description: The IPv4 address of the destination machine
          example: 172.16.3.167
      requestBody:
        description: The operation requires a threshold parameters to be set into
          the machine. The threshold must be between 0 and 100
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/containers_address_body_1'
        required: true
      responses:
        "200":
          description: Correct Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CorrectResponse'
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadResponse'
      x-openapi-router-controller: swagger_server.controllers.default_controller
    delete:
      summary: Remove a registered docker host into the service
      description: Uninstall the selected manager from the machine. The system has
        already stored the password but its required in order to secure the operation
        and prevent that users will disable the managers
      operationId: remove_docker_manager
      parameters:
      - name: address
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: string
          description: The IPv4 address of the selected machine
          example: 172.16.3.167
      responses:
        "200":
          description: Correct Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CorrectResponse'
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadResponse'
      x-openapi-router-controller: swagger_server.controllers.default_controller
    x-temp:
      summary: Remove a registered docker host into the service
      description: Uninstall the selected manager from the machine. The system has
        already stored the password but its required in order to secure the operation
        and prevent that users will disable the managers
      parameters:
      - name: address
        in: path
        required: true
        schema:
          type: string
        description: The IPv4 address of the selected machine
        minLength: ""
      operationId: remove_docker_manager
      responses:
        "200":
          description: Correct Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CorrectResponse'
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadResponse'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                password:
                  type: string
                  example: root
        description: The operation requires the password of the root user on the remote
          machine. It also requires that on the remote machine an ssh connection is
          present on the default port 23
...
components:
  schemas:
    CorrectResponse:
      properties:
        command:
          type: string
          description: "type of results: ok or error"
          example: ok
        description:
          type: string
          description: Brief description of the operation result
          example: Request correctly executed
      example:
        description: Request correctly executed
        command: ok
    BadResponse:
      properties:
        command:
          type: string
          description: "type of results: ok or error"
          example: ok
        type:
          type: string
          description: short description of the error
          example: INVALID_REQUEST
        description:
          type: string
          description: Brief description of the operation result
          example: Request correctly executed
    containers_body:
      type: object
      properties:
        threshold:
          type: number
          example: 60
    containers_address_body:
      type: object
      properties:
        password:
          type: string
          example: root
    containers_address_body_1:
      type: object
      properties:
        threshold:
          type: integer
          example: 70
...